requirements.txt
# Agent de veille IA — Résumé quotidien Gmail
# Framework: PydanticAI + Anthropic Claude Sonnet 4.6

pydantic-ai[anthropic]>=0.12
google-auth-oauthlib>=1.2.0
google-api-python-client>=2.100.0
google-auth>=2.25.0
python-dotenv>=1.0.0
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
